refactor cln helpers
